squash learnerid and learner files, delete deprecated modules
