+++ src/modrover/strategies/forward_explore.py
@@ -1,7 +1,6 @@
 from typing import Dict, Set
 
-from modrover.learner import Learner
-from modrover.learnerid import LearnerID
+from modrover.learner import Learner, LearnerID
 from modrover.strategies.base import RoverStrategy
